Menu categories: make sort_order actually reorder the list
Cast sort_order to int on save, and add a tie-breaker (ORDER BY sort_order, name) so changing a category's number visibly reorders it even when others share the same value.

+++ admin/menu.php
@@ -63,7 +63,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         $stmt->execute([
             $_POST['name'],
             $_POST['description'],
-            $_POST['sort_order'] ?? 0,
+            (int) ($_POST['sort_order'] ?? 0),
             isset($_POST['allow_composition']) ? 1 : 0,
             $_POST['icon'] ?? 'utensils',
             (int) $_POST['category_id']
